update project structure - move components to main directory
