+++ app/global-error.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Button from './ui/button';
+import Button from '../components/button';
 
 export default function GlobalError({
     error,
